requirements.txt
gensim==3.4.0
lightgbm==2.1.0
nltk==3.2.5
numpy==1.14.0
pandas==0.22.0
Pillow==5.1.0
scikit-learn==0.19.1
tqdm==4.23.0
